requirements.txt
pandas==2.2.2
numpy==1.26.4
python-dateutil==2.9.0.post0
XlsxWriter==3.2.0
